Make this gem work with the new site
There have been a few markup changes to the site, so the old parsing
code wasn't quite right. The modernization brings this up to version
2.0. The gem also now uses Bundler for easier local development.

The `meal` and `neighborhood` options were unused, and have been
removed. They were parsed by the `Runner` and used by the `Scraper`, but
the `Marker`, which printed results, never did anything with them.

Internal changes:

* Convert README to Markdown.
* `Marker`: Stop memoizing results that are only used once. Introduce
  `find` to DRY up the code.
* `Scraper`: Use `casecmp` instead of using regexes for case-insensitive
  comparison; use a heredoc instead of `string +=`
* Introduce a few instance methods here and there.
* Rename `@dump` to `@cache`.